refactor(melody): Moved files to a more logic parent folder.
